requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
django-cors-headers>=4.3.0
